Extract classes to individual files
Extract each class to its own file, so that it's a little easier to see
what's going on.
