+++ lib/wikisnakker.rb
@@ -1,287 +1,3 @@
-require 'wikisnakker/version'
+require 'require_all'
 
-require 'digest/md5'
-require 'open-uri'
-require 'yajl'
-require 'set'
-
-module Wikisnakker
-  class Lookup
-    def self.find(ids)
-      lookup = new(ids)
-      property_lookup = new(lookup.properties)
-      lookup.populate_with(property_lookup)
-      lookup
-    end
-
-    def initialize(*ids)
-      ids = ids.flatten.compact.uniq
-      @used_props = Set.new
-      entities = ids.each_slice(50).map do |id_slice|
-        get(id_slice)[:entities]
-      end
-      @entities = entities.reduce(&:merge) || {}
-    end
-
-    def properties
-      @used_props.to_a.map { |e| "Q#{e}".to_sym }
-    end
-
-    def values
-      @values ||= @entities.values
-    end
-
-    def [](key)
-      @entities[key]
-    end
-
-    def populate_with(properties)
-      each_wikibase_entitiyid(@entities) do |entityid|
-        entityid[:value] = properties["Q#{entityid[:value][:"numeric-id"]}".to_sym]
-      end
-    end
-
-    private
-
-    def get(*ids)
-      query = {
-        action: 'wbgetentities',
-        ids: ids.flatten.join('|'),
-        format: 'json'
-      }
-      url = 'https://www.wikidata.org/w/api.php?' + URI.encode_www_form(query)
-      json = Yajl::Parser.parse(open(url), symbolize_keys: true)
-      save_wikibase_entityids(json)
-      json
-    end
-
-    # If a property is set to another Wikidata article, resolve that
-    # (e.g. set 'gender' to 'male' rather than 'Q6581097')
-    # We don't know yet what that will resolve to, and we don't want
-    # to look them up one by one, so keep track of any entity ids we
-    # encounter and then resolve them later in '#populate_with'.
-    def save_wikibase_entityids(json)
-      each_wikibase_entitiyid(json) do |entityid|
-        @used_props << entityid[:value][:'numeric-id']
-      end
-    end
-
-    def each_wikibase_entitiyid(obj)
-      recurse_proc(obj) do |result|
-        next unless result.is_a?(Hash) && result[:type] == 'wikibase-entityid'.freeze
-        yield(result)
-      end
-    end
-
-    # Recursively calls passed _Proc_ if the parsed data structure is an _Array_ or _Hash_
-    # Taken from the json gem.
-    # @see http://git.io/v4Tf7
-    def recurse_proc(result, &proc)
-      case result
-      when Array
-        result.each { |x| recurse_proc x, &proc }
-        proc.call result
-      when Hash
-        result.each do |x, y|
-          recurse_proc x, &proc
-          recurse_proc y, &proc
-        end
-        proc.call result
-      else
-        proc.call result
-      end
-    end
-  end
-
-  module DynamicProperties
-    PROPERTY_REGEX = /^P\d+s?$/
-
-    def method_missing(method_name, *arguments, &block)
-      return super unless method_name.to_s.match(PROPERTY_REGEX)
-      method_name[-1] == 's' ? [] : nil
-    end
-
-    def respond_to_missing?(method_name, include_private = false)
-      method_name.to_s.match(PROPERTY_REGEX) || super
-    end
-
-    def define_property_method(property, &block)
-      define_singleton_method(property.to_sym, &block)
-    end
-
-    def [](key)
-      __send__(key)
-    end
-  end
-
-  class Item
-    include DynamicProperties
-
-    def self.find(ids)
-      lookup = Lookup.find(ids)
-      data = lookup.values.reject { |d| d[:missing] == '' }
-      inflated = data.map { |rd| new(rd) }
-      ids.is_a?(Array) ? inflated : inflated.first
-    end
-
-    attr_reader :id
-    attr_reader :labels
-    attr_reader :descriptions
-    attr_reader :properties
-    attr_reader :sitelinks
-    attr_reader :all_aliases
-
-    def initialize(raw)
-      @id = raw[:title]
-      @labels = raw[:labels]
-      @descriptions = raw[:descriptions] || {}
-      @all_aliases = raw[:aliases]
-      @properties = raw[:claims].keys
-      @sitelinks = Hash[raw[:sitelinks].map do |key, value|
-        [key.to_sym, Sitelink.new(value)]
-      end]
-      raw[:claims].each do |property_id, claims|
-        define_property_method "#{property_id}s".to_sym do
-          # A claim's rank can be either preferred, normal or deprecated. We sort them by
-          # rank in reverse order because lexicographic ordering happens to work for the
-          # known ranks.
-          claims.map { |c| Claim.new(c) }.group_by(&:rank).sort.reverse.map(&:last).flatten
-        end
-
-        define_property_method property_id do
-          __send__("#{property_id}s").first
-        end
-      end
-    end
-
-    # TODO: have an option that defaults to a different language
-    def to_s
-      labels.key?(:en) ? labels[:en][:value] : @id
-    end
-
-    def label(lang)
-      return nil unless labels.key?(lang.to_sym)
-      labels[lang.to_sym][:value]
-    end
-
-    def description(lang)
-      return nil unless descriptions.key?(lang.to_sym)
-      descriptions[lang.to_sym][:value]
-    end
-
-    def aliases(lang)
-      return [] unless all_aliases.key?(lang.to_sym)
-      all_aliases[lang.to_sym].map { |a| a[:value] }
-    end
-  end
-
-  class Claim
-    def initialize(data)
-      @data = data
-    end
-
-    def value
-      mainsnak.value
-    end
-
-    def to_s
-      mainsnak.value
-    end
-
-    def mainsnak
-      @_mainsnak ||= Snak.new(@data[:mainsnak])
-    end
-
-    def qualifiers
-      Qualifiers.new(@data[:qualifiers])
-    end
-
-    def rank
-      @data[:rank]
-    end
-  end
-
-  class Snak
-    def initialize(snak)
-      @snak = snak
-    end
-
-    # https://www.wikidata.org/wiki/Special:ListDatatypes
-    # https://www.wikidata.org/wiki/Help:Data_type
-    def value
-      return if %w(somevalue novalue).include?(@snak[:snaktype])
-      case @snak[:datatype]
-      when 'commonsMedia'
-        # https://commons.wikimedia.org/wiki/Commons:FAQ#What_are_the_strangely_named_components_in_file_paths.3F
-        # commons = 'https://commons.wikimedia.org/wiki/File:%s' % @snak["datavalue"]["value"]
-        val = @snak[:datavalue][:value].tr(' ', '_')
-        md5 = Digest::MD5.hexdigest val
-        "https://upload.wikimedia.org/wikipedia/commons/#{md5[0]}/#{md5[0..1]}/#{val}"
-      when 'wikibase-item'
-        # "Q%s" % @snak["datavalue"]["value"]["numeric-id"]
-        Item.new(@snak[:datavalue][:value])
-      when 'string'
-        @snak[:datavalue][:value]
-      when 'external-id'
-        @snak[:datavalue][:value]
-      when 'quantity'
-        if @snak[:datavalue][:value][:upperBound] == @snak[:datavalue][:value][:lowerBound]
-          @snak[:datavalue][:value][:amount].to_i
-        else
-          warn "FIXME: Unhandled 'quantity': #{@snak[:datavalue][:value]}"
-        end
-      when 'time'
-        case @snak[:datavalue][:value][:precision]
-        when 11
-          @snak[:datavalue][:value][:time][1..10]
-        when 10
-          @snak[:datavalue][:value][:time][1..7]
-        when 9
-          @snak[:datavalue][:value][:time][1..4]
-        when 7
-          '' # Just ignore dates with century precision
-        else
-          warn "FIXME: Unhandled 'time' precision: #{@snak[:datavalue][:value][:precision]}"
-        end
-      when 'url'
-        @snak[:datavalue][:value]
-      else
-        warn "FIXME: '#{@snak[:datatype]}' is not implemented yet in Wikisnakker::Snak#value. Defaulting to empty string. #{@snak[:datavalue][:value]}"
-        ''
-      end
-    end
-  end
-
-  class Sitelink
-    attr_reader :site
-    attr_reader :title
-    attr_reader :badges
-
-    def initialize(raw)
-      @site = raw[:site]
-      @title = raw[:title]
-      @badges = raw[:badges]
-    end
-  end
-
-  class Qualifiers
-    include DynamicProperties
-
-    attr_reader :snaks
-    attr_reader :properties
-
-    def initialize(qualifier_snaks)
-      qualifier_snaks ||= {}
-      @properties = qualifier_snaks.keys
-      qualifier_snaks.each do |property_id, snaks|
-        define_property_method "#{property_id}s".to_sym do
-          snaks.map { |s| Snak.new(s) }
-        end
-
-        define_property_method property_id do
-          __send__("#{property_id}s").first
-        end
-      end
-    end
-  end
-end
+require_rel '.'
